SHARED-86945 fixed charset in request body
(Request.java)
problem: without explicit charset in getBytes() a default JVM charset is taken.
The default charset is not always the same as expected one.

solution:
try to detect a charset from request headers. If the charset is
provided in headers, use the provided one. If not, use default UTF-8
charset.